Replace batchProgressPrinter with taskExecTUI
